Moved image translation to its own file
